Add commands for inspect, repair, upgrade
