+++ PotatoChipMine.Core/Commands/UpgradeCommand.cs
@@ -1,0 +1,30 @@
+using PotatoChipMine.Core.GameEngine;
+using PotatoChipMine.Core.GameRooms.ControlRoom.Services;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace PotatoChipMine.Core.Commands
+{
+    public class UpgradeCommand : CommandWithGameState, ICommand
+    {
+
+    }
+
+    public class UpgradeCommandHandler : ICommandHandler<UpgradeCommand>
+    {
+        public void Handle(UpgradeCommand command)
+        {
+            var gameState = command.GameState;
+
+            var scene = Scene.Create(new List<IGameEntity>
+                {
+                    new UpgradeHandlerEntity(gameState)
+                });
+
+            gameState.PromptText = "Enter Digger Name: ";
+            Game.PushScene(scene);
+            return;
+        }
+    }
+}
