+++ PotatoChipMine.Core/GameEngine/Scene.cs
@@ -4,16 +4,20 @@ namespace PotatoChipMine.Core.GameEngine
 {
     public class Scene
     {
-        public IList<IGameEntity> Entities { get; }
+        public IReadOnlyList<IGameEntity> Entities { get; }
 
-        private Scene(IList<IGameEntity> entities)
+        private Scene(List<IGameEntity> entities)
         {
             Entities = entities;
         }
 
-        public static Scene Create(IList<IGameEntity> entities)
+        public static Scene Create(List<IGameEntity> entities)
         {
             return new Scene(entities);
+        }
+        public static Scene Create(IGameEntity entities)
+        {
+            return new Scene(new List<IGameEntity> { entities });
         }
     }
 }
